Backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0



langchain>=0.1.0
langgraph>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.2



biopython>=1.83          # For PubMed (Bio.Entrez)
fuzzywuzzy>=0.18.0       # For Jan Aushadhi matching
python-Levenshtein>=0.23.0 # Speedup for fuzzywuzzy



python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9



requests>=2.31.0
pydantic>=2.0.0